test(pg): the graph derivation handler live test requires the test database instead of skipping (#878)
The file read OPENBRAIN_TEST_DATABASE_URL itself and swapped in describe.skip
when it was unset, so an absent variable produced 0 pass / 8 skip / 0 fail and
exit 0 -- indistinguishable at the exit code from a suite that ran and passed.
It now calls requireTestDatabaseUrl() at module scope and fails loudly with
test_database_required.

Split by subject at the same time, because the pre-commit gate lints staged
files whole and the 770-line original carried eight findings:

- src/graph-derivation-handler.live.test.ts keeps selection and convergence.
- src/graph-derivation-handler-isolation.live.test.ts takes isolation and
  atomicity, and poolFailingAfter moves with it.
- src/graph-derivation-handler-test-helpers.ts holds the namespaces, auth,
  cleanup, insertSource and jobFor. It creates no pool: each database function
  takes the caller's pool first, so each half ends exactly its own connection.

Test names, order and assertions are unchanged. Each it body is a hoisted
module-scope async function so no describe callback exceeds the code-line
standard, and the five non-null assertions became an expectDefined guard
exported from the helper module.

Both new describe names are registered in scripts/assert-db-tests-ran.ts at
their JUnit-measured counts (4 and 3), with the total floor moved 285 -> 292.

Refs #878
